Rename logsComponent to logDisplayComponent
