+++ src/components/logdisplay/LogDisplayComponent.vue
